move to TokenType.Serializer
